Change: NPM version to 4
The bump in major version is likely not necessary as there are no
truly backwards incompatible changes with the introduction of A+
compliance. However, the resolution behaviour of deferreds _did_
change which makes me want to bump the major version in the off
chance that something relies on, for example, the resolution _not_
handling other thenables.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>