Add: Support for resolving with multiple values
This is intended to help with converting Node.js style functions
that accept callbacks. Sometimes they execute callbacks with
multiple parameters.

The default resolution logic remains unchanged and the package
still passes the upstream specification tests. However, if a
Deferred is resolved with multiple arguments those are now proxied
to the callbacks/errbacks.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>