Fix: Bug causing early resolution of All
The recent change that preserved the order of values in the
results from an All collection introduced a bug where the All
would resolve as soon as the right-most element was ready regardless
of whether or not the other values were ready.

This bug is patched by looking at how many values have been resolved
rather than the position of any given element in the results list.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>